backend/requirements.txt
# Django Core
Django>=5.0,<5.1
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.3,<6.0

# Database
mysqlclient>=2.2,<3.0

# CORS
django-cors-headers>=4.3,<5.0

# API Documentation
drf-spectacular>=0.27,<1.0

# Environment variables
python-dotenv>=1.0,<2.0

# Image handling
Pillow>=10.0,<11.0

# Filtering
django-filter>=23.5,<24.0

# Utilities
python-dateutil>=2.8,<3.0

# For geolocation calculations (Haversine)
haversine>=2.8,<3.0
